Break the module script into a header and a source script.
